Updated Shopping List Test
