+++ recipe2_test
@@ -1,0 +1,4 @@
+name,culture,calories
+Taro Ice Cream Recipe,american,3523.09
+Orange Crush Cake,american,8008.38
+Carrot Cake Smoothie,american,456.793
