+++ src/interface_adapter/shopping_list/SLPresenter.java
@@ -1,7 +1,6 @@
 package interface_adapter.shopping_list;
 
 import interface_adapter.ViewManagerModel;
-import interface_adapter.add_to_collection.AddCollectionViewModel;
 import use_case.shopping_list.SLOutputBoundary;
 import use_case.shopping_list.SLOutputData;
 
@@ -9,17 +8,17 @@ public class SLPresenter implements SLOutputBoundary {
     private ViewManagerModel viewManagerModel;
     private SLViewModel slViewModel;
 
-    private AddCollectionViewModel addCollectionViewModel;
+    //private AddCollectionViewModel addCollectionViewModel;
 
-    public SLPresenter(ViewManagerModel viewManagerModel, SLViewModel slViewModel, AddCollectionViewModel addCollectionViewModel){ //TODO: update to work with current view model (ViewCollection?)
+    public SLPresenter(ViewManagerModel viewManagerModel, SLViewModel slViewModel){
         this.viewManagerModel = viewManagerModel;
         this.slViewModel = slViewModel;
-        this.addCollectionViewModel = addCollectionViewModel;
     }
 
     public SLPresenter() {
 
     }
+
 
     @Override
     public void prepareFailView(String error) {
