+++ test/SLInteractorTest.java
@@ -7,16 +7,14 @@ import interface_adapter.ViewManagerModel;
 import interface_adapter.shopping_list.SLPresenter;
 import interface_adapter.shopping_list.SLState;
 import interface_adapter.shopping_list.SLViewModel;
-import org.junit.Test;
 import use_case.shopping_list.SLDataAccessInterface;
 import use_case.shopping_list.SLInputData;
 import use_case.shopping_list.SLInteractor;
 
-import java.io.*;
+import java.io.IOException;
 import java.util.ArrayList;
-import java.util.stream.Stream;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 public class SLInteractorTest {
 
     private RecipeCollection recipeCollection = new RecipeCollection();
@@ -89,131 +87,101 @@ public class SLInteractorTest {
         recipeCollection.addRecipe(r3);
     }
 
-    @org.junit.Test
-    public void twoIngredientSLTest() throws FileNotFoundException { //1 recipe, 2 items
-        addRecipe2Ingredients();
-        SLInputData slInputData = new SLInputData(recipeCollection);
-        SLDataAccessInterface slDAO = new ShoppingListDataAccessObject("./ShoppingList2.txt");
-        ShoppingListFactory slFactory = new ShoppingListFactory();
-        SLViewModel slViewModel = new SLViewModel();
-        ViewManagerModel viewManagerModel = new ViewManagerModel();
-        SLPresenter slPresenter = new SLPresenter(viewManagerModel, slViewModel);
-        SLInteractor interactor = new SLInteractor(slDAO, slPresenter, slFactory);
 
-        try {
-            interactor.execute(slInputData);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
 
-        assertEquals("shopping list", viewManagerModel.getActiveView()); //sees if view is changed
-        assertTrue(new File("./ShoppingList2.txt").exists());
+//    @Test
+//    public void twoRecipe4SLTest(){ //2 recipes, 5 ingredients (no repeat)
+//        addRecipe2Ingredients();
+//        addRecipe3Ingredients();
+//        SLInputData slInputData = new SLInputData("./recipe");
+//        SLDataAccessInterface slDAO = new ShoppingListDataAccessObject("./ShoppingListTest.txt");
+//        ShoppingListFactory slFactory = new ShoppingListFactory();
+//        SLViewModel slViewModel = new SLViewModel();
+//        ViewManagerModel viewManagerModel = new ViewManagerModel();
+//        SLPresenter slPresenter = new SLPresenter(viewManagerModel, slViewModel);
+//        SLInteractor interactor = new SLInteractor(slDAO, slPresenter, slFactory);
+//
+//        try {
+//            interactor.execute(slInputData);
+//        } catch (IOException e) {
+//            throw new RuntimeException(e);
+//        }
+//
+//        assertEquals("shopping list", viewManagerModel.getActiveView()); //sees if view is changed
+//        assertTrue (new File("./ShoppingListTest.txt").exists());
+//        //check if file exists...
+//    }
 
-        BufferedReader reader = null;
-        try {
-            reader = new BufferedReader(new FileReader("./ShoppingList2.txt"));
-        } catch (FileNotFoundException e) {
-            throw new RuntimeException(e);
-        }
-        Stream<String> line = reader.lines();
+//    @Test
+//    public void twoRecipe11SLTest(){ //2 recipes, 3 + 8 ingredients, 2 ingredient repeats
+//        addRecipe3Ingredients();
+//        addRecipe8Ingredients();
+//        SLInputData slInputData = new SLInputData(recipeCollection);
+//        SLDataAccessInterface slDAO = new ShoppingListDataAccessObject("./ShoppingListTest.txt");
+//        ShoppingListFactory slFactory = new ShoppingListFactory();
+//        SLViewModel slViewModel = new SLViewModel();
+//        ViewManagerModel viewManagerModel = new ViewManagerModel();
+//        SLPresenter slPresenter = new SLPresenter(viewManagerModel, slViewModel);
+//        SLInteractor interactor = new SLInteractor(slDAO, slPresenter, slFactory);
+//
+//        try {
+//            interactor.execute(slInputData);
+//        } catch (IOException e) {
+//            throw new RuntimeException(e);
+//        }
+//
+//        assertEquals("shopping list", viewManagerModel.getActiveView()); //sees if view is changed
+//        assertTrue (new File("./ShoppingListTest.txt").exists());
+//
+//        BufferedReader reader = null;
+//        try {
+//            reader = new BufferedReader(new FileReader("./ShoppingListTest.txt"));
+//        } catch (FileNotFoundException e) {
+//            throw new RuntimeException(e);
+//        }
+//        Stream<String> line = reader.lines();
+//
+//        assertEquals((2+3+8-2), line.count()); //2 for header + ---, 1 extra because each ingredientToString gives \n
+//    }
 
-        assertEquals(4, line.count());
-    }
-
-    @Test
-    public void twoRecipe4SLTest(){ //2 recipes, 5 ingredients (no repeat)
-        addRecipe2Ingredients();
-        addRecipe3Ingredients();
-        SLInputData slInputData = new SLInputData(recipeCollection);
-        SLDataAccessInterface slDAO = new ShoppingListDataAccessObject("./ShoppingList2.txt");
-        ShoppingListFactory slFactory = new ShoppingListFactory();
-        SLViewModel slViewModel = new SLViewModel();
-        ViewManagerModel viewManagerModel = new ViewManagerModel();
-        SLPresenter slPresenter = new SLPresenter(viewManagerModel, slViewModel);
-        SLInteractor interactor = new SLInteractor(slDAO, slPresenter, slFactory);
-
-        try {
-            interactor.execute(slInputData);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-
-        assertEquals("shopping list", viewManagerModel.getActiveView()); //sees if view is changed
-        assertTrue (new File("./ShoppingList2.txt").exists());
-        //check if file exists...
-    }
-
-    @Test
-    public void twoRecipe11SLTest(){ //2 recipes, 3 + 8 ingredients, 2 ingredient repeats
-        addRecipe3Ingredients();
-        addRecipe8Ingredients();
-        SLInputData slInputData = new SLInputData(recipeCollection);
-        SLDataAccessInterface slDAO = new ShoppingListDataAccessObject("./ShoppingList2.txt");
-        ShoppingListFactory slFactory = new ShoppingListFactory();
-        SLViewModel slViewModel = new SLViewModel();
-        ViewManagerModel viewManagerModel = new ViewManagerModel();
-        SLPresenter slPresenter = new SLPresenter(viewManagerModel, slViewModel);
-        SLInteractor interactor = new SLInteractor(slDAO, slPresenter, slFactory);
-
-        try {
-            interactor.execute(slInputData);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-
-        assertEquals("shopping list", viewManagerModel.getActiveView()); //sees if view is changed
-        assertTrue (new File("./ShoppingList2.txt").exists());
-
-        BufferedReader reader = null;
-        try {
-            reader = new BufferedReader(new FileReader("./ShoppingList2.txt"));
-        } catch (FileNotFoundException e) {
-            throw new RuntimeException(e);
-        }
-        Stream<String> line = reader.lines();
-
-        assertEquals((2+3+8-2), line.count()); //2 for header + ---, 1 extra because each ingredientToString gives \n
-    }
-
-    @Test
-    public void sameIngredientDifferentMeasurement(){ //3+3 ingredients (2 with different measurements), 2 recipes
-        addRecipe3Ingredients();
-        addRecipeUnusualMeasurement();
-        SLInputData slInputData = new SLInputData(recipeCollection);
-        SLDataAccessInterface slDAO = new ShoppingListDataAccessObject("./ShoppingList2.txt");
-        ShoppingListFactory slFactory = new ShoppingListFactory();
-        SLViewModel slViewModel = new SLViewModel();
-        ViewManagerModel viewManagerModel = new ViewManagerModel();
-        SLPresenter slPresenter = new SLPresenter(viewManagerModel, slViewModel);
-        SLInteractor interactor = new SLInteractor(slDAO, slPresenter, slFactory);
-
-        try {
-            interactor.execute(slInputData);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-
-        assertEquals("shopping list", viewManagerModel.getActiveView()); //sees if view is changed
-        assertTrue (new File("./ShoppingList2.txt").exists());
-        assertEquals("shopping list", viewManagerModel.getActiveView()); //sees if view is changed
-        assertTrue (new File("./ShoppingList2.txt").exists());
-
-        BufferedReader reader = null;
-        try {
-            reader = new BufferedReader(new FileReader("./ShoppingList2.txt"));
-        } catch (FileNotFoundException e) {
-            throw new RuntimeException(e);
-        }
-        Stream<String> line = reader.lines();
-
-        assertEquals((2+6), line.count()); //2 for header + ---, 1 extra because each ingredientToString gives \n
-    }
+//    @Test
+//    public void sameIngredientDifferentMeasurement(){ //3+3 ingredients (2 with different measurements), 2 recipes
+//        addRecipe3Ingredients();
+//        addRecipeUnusualMeasurement();
+//        SLInputData slInputData = new SLInputData(recipeCollection);
+//        SLDataAccessInterface slDAO = new ShoppingListDataAccessObject("./ShoppingListTest.txt");
+//        ShoppingListFactory slFactory = new ShoppingListFactory();
+//        SLViewModel slViewModel = new SLViewModel();
+//        ViewManagerModel viewManagerModel = new ViewManagerModel();
+//        SLPresenter slPresenter = new SLPresenter(viewManagerModel, slViewModel);
+//        SLInteractor interactor = new SLInteractor(slDAO, slPresenter, slFactory);
+//
+//        try {
+//            interactor.execute(slInputData);
+//        } catch (IOException e) {
+//            throw new RuntimeException(e);
+//        }
+//
+//        assertEquals("shopping list", viewManagerModel.getActiveView()); //sees if view is changed
+//        assertTrue (new File("./ShoppingListTest.txt").exists());
+//        assertEquals("shopping list", viewManagerModel.getActiveView()); //sees if view is changed
+//        assertTrue (new File("./ShoppingListTest.txt").exists());
+//
+//        BufferedReader reader = null;
+//        try {
+//            reader = new BufferedReader(new FileReader("./ShoppingListTest.txt"));
+//        } catch (FileNotFoundException e) {
+//            throw new RuntimeException(e);
+//        }
+//        Stream<String> line = reader.lines();
+//
+//        assertEquals((2+6), line.count()); //2 for header + ---, 1 extra because each ingredientToString gives \n
+//    }
 
     @org.junit.Test
     public void testEmptyRecipeCollection(){ //emptySL
-        RecipeCollection recipeCollectionEmpty = new RecipeCollection();
-        SLInputData slInputData = new SLInputData(recipeCollectionEmpty);
-        SLDataAccessInterface slDAO = new ShoppingListDataAccessObject("./ShoppingList2.txt");
+        SLInputData slInputData = new SLInputData("./recipe_empty_test");
+        SLDataAccessInterface slDAO = new ShoppingListDataAccessObject("./ShoppingListTest.txt");
         ShoppingListFactory slFactory = new ShoppingListFactory();
         SLViewModel slViewModel = new SLViewModel();
         ViewManagerModel viewManagerModel = new ViewManagerModel();
